requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==39.0.0
distlib==0.3.6
docopt==0.6.2
dulwich==0.20.50
filelock==3.9.0
gTTS==2.3.1
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
Js2Py==0.74
jsonschema==4.17.3
keyboard==0.13.5
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.0.0
MouseInfo==0.1.3
msgpack==1.0.4
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pexpect==4.8.0
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.6.2
playsound==1.3.0
poetry-core==1.4.0
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycparser==2.21
pygame==2.1.3.dev8
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyPrind==2.11.3
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
pySmartDL==1.3.4
pytesseract==0.3.10
python3-xlib==0.15
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
scipy==1.10.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sounddevice==0.4.5
soundfile==0.11.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
tk==0.1.0
toml==0.10.2
tomlkit==0.11.6
trove-classifiers==2023.1.20
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
virtualenv==20.17.1
wavio==0.0.7
webencodings==0.5.1
zipp==3.11.0
